Improve vvcs#remote#execute() failure indication
The changes from the last commit had the side-effect delaying the abortion of
a comparison when remote commands failed. In order to fix this the failure
indication from vvcs#remote#execute() was improved.

The test setup was changed to parse the stub keys in a determined order (the
keys() method return them in unspecified order), in order to avoid overly
complex regexes.